simple_webserver/requirements.txt
asgiref==3.5.0
click==8.1.2
Django==4.0.3
Flask==2.1.1
image==1.5.33
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.0.1
six==1.16.0
sqlparse==0.4.2
Werkzeug==2.1.0
zipp==3.7.0
